+++ source/Energinet.DataHub.MarketRoles.EntryPoints.Common/ServiceBusSessionIdMiddleware.cs
@@ -18,7 +18,7 @@ using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Middleware;
-using Processing.Infrastructure.Correlation;
+using Processing.Infrastructure.Configuration.Correlation;
 using Processing.Infrastructure.Serialization;
 
 namespace Energinet.DataHub.MarketRoles.EntryPoints.Common
