+++ source/Processing.Api/Program.cs
@@ -44,13 +44,13 @@ using Processing.Domain.SeedWork;
 using Processing.Infrastructure;
 using Processing.Infrastructure.BusinessRequestProcessing;
 using Processing.Infrastructure.BusinessRequestProcessing.Pipeline;
+using Processing.Infrastructure.Configuration.Correlation;
 using Processing.Infrastructure.Configuration.DataAccess;
 using Processing.Infrastructure.Configuration.DataAccess.AccountingPoints;
 using Processing.Infrastructure.Configuration.DataAccess.Consumers;
 using Processing.Infrastructure.Configuration.DataAccess.EnergySuppliers;
 using Processing.Infrastructure.Configuration.DataAccess.ProcessManagers;
 using Processing.Infrastructure.ContainerExtensions;
-using Processing.Infrastructure.Correlation;
 using Processing.Infrastructure.DomainEventDispatching;
 using Processing.Infrastructure.EDI;
 using Processing.Infrastructure.EDI.ChangeOfSupplier;
